Move eachPackage to packages.forEachIn
