+++ lib/status.js
@@ -3,12 +3,12 @@ var runSync = require("./runsync");
 var path = require("path");
 var OS = require("os");
 
-var eachPackage = require("./eachPackage");
+var packages = require("./packages");
 
 module.exports = function(workspace) {
   var statuses = [];
 
-  eachPackage(workspace, function(packageName, srcPath) {
+  packages.forEachIn(workspace, function(packageName, srcPath) {
     var branchResult = runSync('git', ["symbolic-ref", "HEAD"], srcPath, console, true);
     var branch = branchResult.stdout.toString().replace("refs/heads/", "").trim();
